Db refactor

+++ TechStoreBackend/src/TechStore.Domain/Entities/ProductAggregate/Review.cs
@@ -7,7 +7,7 @@ namespace TechStore.Domain.Entities.ProductAggregate
     {
         public string Email { get; set; }
         
-        public decimal Rate { get; set; } = 0;
+        public int Rate { get; set; } = 0;
         
         public string Comment { get; set; }
